Move csvData & filtering function to top, fix bugs
